requirements.txt
# Core dependencies
python-dotenv
Pillow
reportlab
matplotlib
ttkbootstrap
